Updated 3.7.5
Same Site Cookie Settings - We temporarily disabled sameSite Cookie settings to remove the third party cookie warning.

Havok Condition for Babylon Engine - The Havok code is only available for the Babylon v6.x.x engine. To make sure you do not try to run it with the v5.x.x engine, additional conditional statements have been added. This will keep the Engine and Physics Engine loading properly.

Updated Havok init Code - We updated the Havok physics engine Init Code to include global variables for havokInstance and hk for the Havok plugin. The core file has the init code in the initEnvironment function.

Babylon Engine Settings updated - We updated the settings on the Babylon Engine to disable preserveDrawingBuffer, enable stencil, and disable disableWebGL2Support.

Updated Babylon Engine to Version 6.46.0 - We Updated the Babylon JS Engine to version 6.46.0. This will give us the latest fixes and enhancements available. Note that in your admin, you may need to select version 6 in Settings to use this version.

CSG Rendering Groups - When editing a CSG (Constructive Solid Geometry) objects, the combined object would disappear when editing. It was appearing before the ground layer in the view. Now the rendering group is correct and it is displaying as expected. The image shows then and now as contrast of the edit process shown.

Profile Menu corrected Save Text - On the Profile Menu under the HUD Menu, press M to show, the Save Button was missing the Save Text. The rendering group was not set correctly and caused it to be under the visible layer. It has been corrected.

HUD Menu Rendering Correction - The HUD Menu, press M to show, had an issue with the rendering order of things. This was corrected by adding the Rendering Group ID as 3 to the menu item.