Updated 3.7.6
SError Missing Class Call - There were 4 instances in the classes where the serror function was being called without the class $wtwHandlers-> instance reference. They have been corrected and are now working as expected.

Save Point function Missing ID - The Save Point function was missing the Randomly assigned ID. Without it, it would not save more than one point. It is now functioning as expected.

Add Point Process Corrected - The Add Point function was not opening the point for edit in the form to the right sorted order. The name only has one hyphen and the reference was looking past the end of the name. We corrected it to look for segment 1 (0 based) instead of segment 2, which corrected the issue.

Edit Point Process Corrected - The Edit Point function was not opening the point for edit in the form. The name only has one hyphen and the reference was looking past the end of the name. We corrected it to look for segment 1 (0 based) instead of segment 2, which corrected the issue.

New Tube Default Points - We added code to add 2 points to any new Tube or Line. This makes it easier to get started.

Formatting Numbers - We now have more code to formatting numbers as the edit mold form is opened. This makes the decimal places more consistent.

Default Tube for Open Form for Edit - When you edit a tube and it is loaded into the edit form, if the tube has no points listed, it will default to the new tube 2 points. That way it is shown and will be easier to start editing.

Coordinate Order for Editing Points - Since the points are listed as (x,y,z) in the series, when editing a point, it made more sense to list them in the same order. This is different than the order listed above for the object where we use Z, X, Y.

Edit Point on Click Event - The code was using onmousedown to edit a point. We updated it to the more proper onclick event. It now works as expected.

New Tube Defaults Updated - When adding a new Tube, we now have the defaults set to create a tube with 2 default points, one 10 units over the other point. For Things, the base will be located at (0,0,0) while for Buildings and Communities molds the base is located in front of the camera where you are located. Note that the base coordinate is a reference point for the object, while the tube can use any point values required.

Edit Mold Decimal Points - When editing a mold, or adding a new mold, the position, rotation, and scaling were not displaying the same number of decimal points until changed. We added the function formatNumber to the values on opening and now they all have 2 decimal places.

Wireframe Settings in the Complete Mold Function - The completeMold function required some direct settings for the wireframe that excludes other potential settings used on other meshes.

Added Wireframe as a Material Covering - We added the wireframe as a material for a covering of a mesh. It is now on the drop down list for coverings when you add a mesh.

Path Number reference was incorrect - In the get Mold Points function in the wtwdb class, the reference to $zpathnumber was missing the z. It has been corrected so that it will now save the path points.

Camera Distance with Meshes without Collisions - When you walk through an object (like a partially transparent box) that has collisions turned off, we did not want the camera distance to be shortened to the box as you enter it. This change checks if the mesh has collisions and if the opacity is less than one before adjusting the distance.