Updated 3.6.7
Copy 3D Thing Function Corrected - The Copy 3D Thing function was not working because of 3 invalid references. The first was the use of variables for diffusecolor, specularcolor, emissivecolor, and ambientcolor instead of the $zrow object. The second was not handling the possibility for null values passed as the unloadactionzoneid, attachactionzoneid, and altloadactionzoneid (which have also been corrected in the Copy Building and Copy Community functions). The third was a reference to moldid when the select statement was returning thingmoldid as the name referenced. We updated the select statement to use moldid.