Updated 3.7.6
Expanded the Media Library - Instead of nesting the media library menu options, we expanded the items into the top level menu in their own section of the menu, which now includes the 3D Plugins.

New Select 3D Avatar for Edit - We added a new Select Avatar page. This page is shown instead of the Admin Menu listing of Avatars to open. This design is similar to the WalkTheWeb Downloads Page and can be set to show between 1 and 4 columns at a time.

3D Webs Listing Page - The 3D Webs listing page was added to the wtwadmin class. This creates the page form layout.

New Select Webs Page - We added a new Select Webs page (Select 3D Community Scene, Select 3D Building, and Select 3D Thing). This page is shown instead of the Admin Menu listing of Webs to open. This design is similar to the WalkTheWeb Downloads Page and can be set to show between 1 and 4 columns at a time.

Admin menu Colors Added - To separate the different 3D Webs, we changed the colors of each so that it becomes more obvious. Green for communities, blue for buildings, yellow for things, and red for avatars. This was done in the wtwadminmenus class, addAdminMenuItem function, which we also added one more parameter for color. Colors accepted are currently limited to the 4 listed above.

Get File List Function Added - We added a getFileList function to the wtwdb class, then extended it to be included in the wtwhandlers, wtwconnect, and wtwplugins classes.

Reordered Admin Menu - We reordered the Admin menu to improve the selections by category. There are now sections for the main components (3D Communities, Buildings, Things, and Avatars), Media Library (files that support the main components), Settings, and Users.

3D Internet Admin Menu - The 3D internet Admin menu was opening an additional menu option in building section. This has been resolved in the class_main.js file adminMenuItemSelected function for the 3D Internet Plugin.

Updated Label for File Types supported - When uploading a 3D Model, the file types supported are currently .babylon, .obj, .glb, and .gltf. THe old label had blender instead of babylon.

Removed Opacity Condition from Camera Distance - We were experimenting with checking the opacity before shortening the camera distance when standing in front of a wall. This was to watch for the difference of a window and a solid wall. But we decided to not use this condition. Now it only checks to see if it checks collisions or not.