Updated 3.7.6
Open List Items Function - The open list items function is used to call and populate the formatted list of available web items from the connect websitems.php file. This funciton is called from the admin menu item.

Webs Items Connect File - There is a new Connect file for websitems.php that is used to create the list of editable items for any given web (3D Community, 3D Building, or 3D Thing). It is only available to the same site it resides on, per the headers passed. It will show Molds, Action Zones, 3D Things, and 3D Buildings added to any 3D Scene.

Proper Case Function - We added a propercase function to help with mix case titles. Previously there was a propercase construct for strings only, but this change makes it more universally available to code.

Added additional trapping to highlighting - We added additional error trapping to the highlight and unhighlight functions. This helps to not generate errors when items are being loaded and unloaded during highlighting stages.

Items to Edit Listing Highlighting - When available, if you hover over an item on the list, it will highlight in the 3D Scene. The item has to be visible in the 3D Scene and fully loaded to apply. Highlighting Child items of a given item are not yet supported.

List All Items to Edit - We added a menu option to List All Items to Edit under the Edit 3D Webs (Communities, Buildings, and Things). This makes it easier to select an item to edit. This also helps when an item is transparent or not selectable in the 3D Scene. We also added Action Zones to the list to make them easier to select and edit.

Mold Positioning with Action Zone Parent - When adding multiple 3D Things with Action Zones (like a swinging door zone) to a 3D Building or 3D Community Scene, the molds were being adjusted for position multiple times. We resolved this issue by checking the parent name of the mold (an action zone) index and making sure the position adjustment to the molds only took place on the same matching instance.

Set New Action Zone Children Position - When editing a swinging door or similar action zone, when you moved the axle position, the children molds would also move. This was due to an incorrect index on the name parts of the molds. When we added local- prefix it shifted the index of the name parts. Now they are set correctly in the setNewActionZone function.

Axle Pole Rotation - When editing a Swinging Door Action Zone, the pole was not reloading to the correct rotation. The rotation of the pole is actually supposed to be zero, since the parent base is what actually is rotated to make the axle take on an angle. It has been corrected in the basic action zones definitions. Note that in production, the pole is not shown; it is only a guide during editing.

Corrected Action Zone Index - When editing a Mold that is attached to an action zone like a swinging door, the index of the action zone was being referenced as the second variable separated by hyphens, but it needed to be the third ever since the local- prefix was added. The index value was updated to be 2 (zero based, third variable). This corrected the ability to move a mold that is parented to an action zone.

Increased Intensity on Images - When you add an Image from 3D Web Objects the image was using the diffuse texture only and was subject to shadows. We changed the covering material to also load the image as the emissive texture to force a brighter image. This is applied by a condition that the name of the mold includes the text mainimage.