Updated 3.7.6
Added Action Zone ID 2 to Molds Tables - On the communitymolds, buildingmolds, and thingmolds tables we added an actionzone2id field that acts as a second actionzoneid field. THe first field was used to group parts in action zones like parts of a door that will swing or slide. The new actionzone2id is used to set the action zone for the mirror (or any other similar type of need). This new field was added to all connect, handlers, classes, forms, and scripts to complete the need and group and save with molds. We also updated the hub to support it in shared items.

Reflective Glass from a Mirror Zone - You can also create reflective glass by setting the opacity level of the mold. You can also set the colors of the mold to tint the glass, solid black (#000000) colors will make the glass mirror even more perfect reflection. The Mirror Zone as a covering will add the reflective objects to the glass.

Mirror Action Zone - We added a Mirror Action Zone. This Zone does not add a mirror, but creates the zone that decides which 3D Objects will be reflected in a mirror. After adding the zone, add a 3D Object and set the Covering to the Mirror Action Zone you added. These two parts create a mirror and sets all objects to add to the reflection list. Note that when an avatar enters the zone, it is automatically added to the reflection list.

Mirror as a Covering (Mesh Material) - The Mirror Covering is actually added by a Mirror Action Zone. By Adding a Mirror Action Zone on your 3D Scene, any 3D Object in the Mirror Zone will be added to the reflection list for the Mirror Covering. Using this method, you can create different zones of objects and add the mirror covering as reflecting any object in that zone. This covering can be used on most shapes. To get started, add the Mirror Zone from Action Zones, it will then show up as an option under the covering drop-down when editing a shape.

Upgraded to Babylon JS Version 7! - We upgraded the Babylon default version to 7.19 and set the defaults to v7.x.x. In admin you can manually select an engine version under Settings, Server Settings.

Showing Action Zones - When in Admin Mode and show Action Zones is on in the quick edit menu, if the action zone was not loaded at the time the quick menu was set, it would not show when loaded. We corrected this condition by updating the processMoldQueue function to check for the button setting when a new action zone is loaded and show the action zone as needed. This change will not activate in Browse mode, only Admin Mode.

Action Zone Render Group - When in Admin Mode and editing an Action Zone, we added the setting to the showActionZone function to set the render group to 1. This will make them visible in the correct layer during editing.

Corrected the Edit Pole Height - When editing an action zone with an axis, the pole that defines the axis was disappearing for some types of zone. We corrected the pole dimensions to make sure it is visible as expected.

Click Open Swinging Doors - We now have click open swinging doors. The click open doors require your avatar to be inside the zone then click to open the door. The door will swing shut when your avatar leaves the zone.

Click Open Sliding Doors - We now have click open sliding doors. Formerly called click activated, the click open doors require your avatar to be inside the zone then click to open the door. The door will slide shut when your avatar leaves the zone.

New Sliding Door Rotation - When adding a new Sliding Door, the axle was 90 degrees off from the avatar position. It has now been corrected to face correctly.

Action Zone Form Decimal Places - When adding a new Action Zone, some fields were not set to show 2 decimal places. Now they are all consistent.

Click Mold added Parent for Conditions - When you click on a mold, the code at inputClick in wtw_inputs.js now identifies the parent. Using the parent you can test to see if the mold is parented to an action zone. In this instance, we test to see if it is a click open action zone. If it is, we test if the avatar is in the action zone then open the door.

Click Move Disabled if not Avatar Camera - The Click to move the avatar has been disabled when in Admin Mode and if the Avatar Camera is not active. This allows you to work with the scene easier to click items for edit without moving the avatar.

Corrected Coins Render Group - The coins were being rendered behind the ground and other objects. We corrected the render group to 1 and now it is part of the same render layer as the other items in the 3D Scene. This corrects the missing coins in scenes.

Edit 3D Model Enhancements - Overall when editing a 3D Model we made a number of improvements; widened the text box showing the selected file, fixed the guidelines to show the full dimensions of the 3D Model, and fixed the ability to highlight all parts of the 3D Model.

Show and Hide Action Zones on List Hover - When listing all Items in a 3D Web, the action zones now will show and hide as the list item is hovered. This will make it easier to find and identify them in the 3D Scene.

New Set Highlight Recursive Function - We added a new setHighlight function that is called by highlightMold and unhighlightMold functions that will now check all children parts for highlighting and removing the highlight. This new function supports child meshes of 3D Models, standard meshes, and complete 3D Webs in a 3D Scene (like a building in a community or thing in the building or community).

Get Mesh or Node By ID - Many objects are parent based on meshes or transform node. this function checks for either when necessary and allows us to check the child objects regardless of the type. We now replaced the use of scene.getTransformNodeByID with WTW.getMeshOrNodeByID in all code.

Edit Guide Lines for 3D Webs - We added support to scale up the guide lines when editing a 3D Thing or 3D Building within a 3D Web Scene. Previously the guide lines were only showing the scale of the base.

Edit Poles Bounding Box Info - The openEditPoles function now checks for the WTW.boundingbox information to determine the guide locations. It also checks the total length of the item and appends the length of the guides to make sure they extend past the object being edited.

Added Bounding Tracking Info to 3D Molds - To assist with creating the bounding box in Admin mode, we append the 3D Model def with bounding information min and max points. This keeps it a one time calculation on creation and reusable each time it is edited. Production Browse mode does not create this extra information.

Enhanced Bounding Box on 3D Models - Up until now, when editing a 3D Model, the guide lines were focused on the size of the parent base box. Now we calculate the bounding box for the 3D Model and extend the guide lines to the outer extremities of the 3D Model. This will help in aligning items in a 3D Scene.