Updated 3.8.0
Cast Shadows Setting - When editing a mold, under Advanced Settings, there are two options that effect shadows; Allow Shadows on Surface and Allow to Cast Shadows. By default these are not set. Each item you select adds to the workload to generate a smooth scene. You may need to balance what is included with the ability to keep smooth animation, especially in larger scenes.

Cast Shadows Field Added - For every instance of recieveshadows in relation to molds (communitymolds, buildingmolds, and thingmolds) we added the field for castshadows. This field will allow us to select which objects can cast a shadow. The reason for this field instead of just allowing all shadows from all objects, is that some large objects may distort the resolution of the shadow map. When this happens, the shadow may appear as one large rectangle instead of the detailed pattern based on the object. This setting allows you to test objects shadows and select which ones are most effective in your scenes. Note that a 3D Model cast shadows setting affects the full model. To change this to a more selective shadow pattern, you may need to separate the 3D Model into multiple models.

Communities Sun and Back Light Positions - We added fields for the sunpositionx, sunpositiony, sunpositionz, backlightpositionx, backlightpositiony, and backlightpositionz. These values propagate through all levels of code including: connect, handlers, class functions, forms, and JavaScripts. Everywhere the sundirection or backlightdirection are called, the position fields were added. Even though directional light does not use the position, we added it to the fields so we can test various light settings present and future. Note that the back light currently is set to use a point light which can use position.

Avatar Shadows Engaged - When we add an avatar, we now have the initial shadows enabled at the time of avatar creation. This also includes the ability to receive shadows on the avatar meshes.

Form Field for Z3 - A number of form fields using position were mislabeled where positionx3 was duplicated instead of showing positionz3 for the z axis position. These are all corrected in the code.

Set Sun Light Function Updates - We fixed the setSunLight function to load the proper values for adjustable lighting. It now works in admin when you adjust the values. Note that the sunlight and the back light are at values near 1000 (positive or negative). Closer values tend to show a strong light reflection on the ground. This can also be adjusted in the light color settings.

New Reload Shadows Function - We added a new function called reloadShadows that will check the existing molds for the castshadows flag and add them to the newly created shadow generator. This is most helpful when in admin and changing settings.

Updated Set Shadow Settings Function - The setShadowSettings function is used to initiate the shadow generator and apply the initial settings. We have streamlined the used shadow options. Note that there are many more options not in use and yet some are referenced in this section of code.

Added new Edit Lighting Settings - We redesigned the edit lighting settings form. In general, we have Sun light from one direction, then a back light from the opposite above angle to give the objects a hint of light in the shadows. This helps differentiate between day and night shadows in a scene.

Edit Profile added to Open Login HUD - The openLoginHUD function now supports the 3D Mold for the Edit Profile 3D Form, all of its button animations, and load border focus on the fields.

HUD Login Click Function Updated - The hudLoginClick function was updated to now support the new Edit Profile 3D Form, including: Open the Edit Profile 3D Form from the User Menu 3D Form, Save Profile, and X to close the form. Furthermore, the X to close the 3D Forms has been added to all of the HUD Login Menus and Forms.

Change Login HUD Focus Field - The changeLoginHUDFocus function basically grays all of the form field borders then changes the one form field receiving focus to have a selected field color. The additional field names required for the edit profile form had to be added to this function.

HUD Login Focus Text Form Controls Updated - In the hudLoginFocusText function, the form fields for the Edit Profile 3D Form were added to the tab order listing and the offset for each text entry was also added to the second part of the function that helps with the alignment of the text entered into the form fields.

HUD Login Load Profile Function Added - A new function was added for the HUD Login (3D Forms) to load the profile when the form is opened; hudLoginLoadProfile.

HUD Login Save Profile - A new function was added for the HUD Login (3D Forms) to save the profile; hudLoginSaveProfile. This uses the saveMyProfile wtwusers Class function.

Save My Profile Class Function Updated - In the saveMyProfile function in the wtwusers class, we added code to the function to also update the useravatars table display name. This will make sure the current active avatar receives the display name. Prior to this, only the default display name was updated in the users table.

Save My Profile Update - In the handler, user class, and user JavaScript functions, we updated the code to also pass the user avatar id. This will assist in updating the current avatar display name as well as the default display name under the user.

New Edit Profile 3D Form - The Edit Profile page has now been converted to a 3D Form. This also makes it accessible in-game from VR and AR.

New Local Checkbox on Login Page - We replaced the Login Menu that showed buttons for Global WTW Login, Local Login, and Guest options with the login form with a checkbox that switches the form from Global to Local. Also, after you login, it will automatically load your last used avatar (as long as you allowed cookies to save the settings). This cuts out many steps on login and getting started on a new 3D Scene.

New Startup Sequence - We hid the Enter Button and automatically start the user with a generic robot. We still show the Login button for anonymous users. This prompts them to login or provides options to create an account.

Select Avatar Close X Button - We added a close X button on the top right of the Select An Avatar 3D Form.

Login Menu Close Button - We added a - X - Close button to the Login Menu.

Edit Avatar Loading Sequence Update - There was code in the core script file that checked for Admin View to load the Avatar being edited. The previous reason for adding it there has been removed so it has now been relocated to the Admin Init script file with the other webs checking and loading.

Corrected Copy Thing Function - The Copy Thing function in the things class was not updating the actionzoneid and actionzone2id. This was causing action zones to not fully function after the copy. It is now modeled after the copy building function and works as expected.

Corrected Copy Thing - In Admin Mode, Copy 3D Thing, the object was not copying the thingmolds. It was missing the new actionzone2id field. It has been added to correct the issue.